feat: agregar métodos CRUD para Paciente, Odontólogo y Domicilio y manejo de logs con logger

+++ src/main/java/com/example/ProyectoIntegradorPrueba/service/OdontologoService.java
@@ -3,6 +3,8 @@ package com.example.ProyectoIntegradorPrueba.service;
 import com.example.ProyectoIntegradorPrueba.model.Odontologo;
 import com.example.ProyectoIntegradorPrueba.repository.IOdontologoRepository;
 import jakarta.persistence.EntityNotFoundException;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,34 +14,64 @@ import java.util.Set;
 
 @Service
 public class OdontologoService implements IOdontologoService{
-
+    private static final Logger logger = LogManager.getLogger(OdontologoService.class);
     @Autowired
     private IOdontologoRepository iodontologoRepository;
 
     @Override
-    public void guardarOdontologo(Odontologo odontologo) {
-        iodontologoRepository.save(odontologo);
+    public Odontologo guardarOdontologo(Odontologo odontologo) {
+            //funcion para evitar duplicados utilizando la matricula
+            logger.info("Iniciando guardarOdontologo con datos: {}", odontologo);
+          return iodontologoRepository.save(odontologo);
     }
 
     @Override
     public Odontologo buscarPorId(Long id) {
+        // findById carga la entidad completa
+        logger.info("Iniciando buscar odontologo con: {}", id);
         Optional<Odontologo> odontologoBuscado = iodontologoRepository.findById(id);
         if (odontologoBuscado.isPresent()) {
             return odontologoBuscado.get();
         } else {
-            return null;
+            throw new EntityNotFoundException("El odontólogo con el ID " + id + " no existe.");
         }
     }
 
     @Override
-    public void modificarOdontologo(Odontologo odontologo) {
+    public Odontologo buscarPorMatricula(String matricula) {
+        logger.info("Iniciando buscar odontologo por matricula: {}", matricula);
+        Optional<Odontologo> odontologoBuscadoByMatricula = iodontologoRepository.findByMatricula(matricula);
+        if (odontologoBuscadoByMatricula.isPresent()) {
+            return odontologoBuscadoByMatricula.get();
+        } else {
+            throw new EntityNotFoundException("El odontólogo con la matricula " + matricula + " no existe.");
+        }
+    }
 
+    @Override
+    public Odontologo buscarPorNombreApellido(String nombre, String apellido) {
+        logger.info("Iniciando buscar odontologo: {}", nombre, apellido);
+        String nombreNormalizado = nombre.toLowerCase();
+        String apellidoNormalizado = apellido.toLowerCase();
+        Optional<Odontologo> odontologoBuscadoByNombreAndApellido = iodontologoRepository.findByNombreAndApellido(nombreNormalizado, apellidoNormalizado);
+        if (odontologoBuscadoByNombreAndApellido.isPresent()) {
+            return odontologoBuscadoByNombreAndApellido.get();
+        } else {
+            throw new EntityNotFoundException("El odontólogo" + nombre+ " "+apellido + " no existe.");
+        }
+    }
+
+
+    @Override
+    public Odontologo modificarOdontologo(Odontologo odontologo) {
+        logger.info("Iniciando modificacion de odontologo: {}", odontologo);
         //iodontologoRepository.save(odontologo);
         Optional<Odontologo> odontologoExistente = iodontologoRepository.findById(odontologo.getId());
 
         if (odontologoExistente.isPresent()) {
             // Si existe, se actualizan los datos
             iodontologoRepository.save(odontologo);
+            return odontologoExistente.get();
         } else {
             // Si no existe, puedes manejar el caso según sea necesario
             // Por ejemplo, lanzar una excepción o retornar un mensaje de error
@@ -49,7 +81,12 @@ public class OdontologoService implements IOdontologoService{
 
     @Override
     public void eliminarOdontologo(Long id) {
-       iodontologoRepository.deleteById(id);
+        logger.info("Iniciando eliminacion de odontologo: {}", id);
+       if(iodontologoRepository.existsById(id)) {
+           iodontologoRepository.deleteById(id);
+       }else {
+           throw new EntityNotFoundException("El odontólogo con el ID " + id + " no existe.");
+       }
     }
 
     @Override
